+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 0.0.5
 - added cart support with shipping, items (partial), promos, customer - check_cart
+- added order table
+- added generic shopgate payment method
 
 # 0.0.4
 - Added authentication configuration