+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 - added generic payment method
 - added generic shipping method
 - added is_shopgate rule for shipping/payment to not appear on frontend
+- added cron cancellations/shipping
+- fixed sort order for categories
+- fixed inactive/invisible categories not exporting
+- fixed products referencing a root category
+- fixed cart qty issues
+- fixed installation routines
 
 # 0.0.4
 - Added authentication configuration